// apps/restaurant-web/src/types/api.ts
export interface Order {
  id: string;
  orderNumber: string;
  customerName: string;
  customerPhone?: string;
  branchId?: string;
  itemsSummary: string;
  totalAmount: number;
  status: string;
  orderType: string;
  createdAt: string;
  driverName?: string;
  driverPhone?: string;
  customerAddress?: string;
  customerLatitude?: number;
  customerLongitude?: number;
  deliveryCost?: number;
}